// apps/web/src/types/api.generated.ts
export interface components {
    schemas: {
// ...
        CreateBranchDto: {
            /** @example فرع المعادي */
            name: string;
            /** @example المعادي */
            area?: string;
            /**
             * @description أنواع المناهج — للمدارس فقط
             * @example [
             *       "LANGUAGES"
             *     ]
             */
            curriculums?: ("ARABIC" | "LANGUAGES" | "IG")[];
            /**
             * @description نوع الملكية — للمدارس فقط: PRIVATE | GOVERNMENT
             * @example PRIVATE
             */
            ownership?: string;
            /** @example false */
            isMain?: boolean;
        };
// ...
    };
    responses: never;
    parameters: never;
    requestBodies: never;
    headers: never;
    pathItems: never;
}